requirements.txt
django
djangorestframework
social-auth-app-django
social-auth-core
GitPython
django-sslserver
python-dateutil
django-ckeditor